+++ src/query_params.ts
@@ -2,6 +2,7 @@ import check from './typecheck';
 import isString from 'lodash/isString';
 import isNumber from 'lodash/isNumber';
 import isPlainObject from 'lodash/isPlainObject';
+import isBoolean from 'lodash/isBoolean';
 
 export const paramValidators = {
     fields: check(
@@ -39,6 +40,26 @@ export const paramValidators = {
     timeZone: check(isString, 'the value for `timeZone` should be a string'),
 
     userLocale: check(isString, 'the value for `userLocale` should be a string'),
+
+    method: check((method): method is 'json' | 'string' => {
+        return isString(method) && ['get', 'post'].includes(method);
+    }, 'the value for `method` should be "get" or "post"'),
+
+    returnFieldsByFieldId: check(
+        isBoolean,
+        'the value for `returnFieldsByFieldId` should be a boolean'
+    ),
+
+    recordMetadata: check(
+        check.isArrayOf(isString),
+        'the value for `recordMetadata` should be an array of strings'
+    ),
+};
+
+export const URL_CHARACTER_LENGTH_LIMIT = 15000;
+
+export const shouldListRecordsParamBePassedAsParameter = (paramName: string): boolean => {
+    return paramName === 'timeZone' || paramName === 'userLocale';
 };
 
 export interface SortParameter<TFields> {
@@ -57,4 +78,7 @@ export interface QueryParams<TFields> {
     cellFormat?: 'json' | 'string';
     timeZone?: string;
     userLocale?: string;
+    method?: string;
+    returnFieldsByFieldId?: boolean;
+    recordMetadata?: string[];
 }