+++ test/select.test.js
@@ -319,7 +319,7 @@ describe('record selection', function() {
         testExpressApp.set('handler override', function(req, res) {
             expect(req.method).toBe('GET');
             expect(req.url).toBe(
-                '/v0/app123/Table?maxRecords=50&sort%5B0%5D%5Bfield%5D=Name&sort%5B0%5D%5Bdirection%5D=desc&cellFormat=json'
+                '/v0/app123/Table?maxRecords=50&sort%5B0%5D%5Bfield%5D=Name&sort%5B0%5D%5Bdirection%5D=desc&cellFormat=json&returnFieldsByFieldId=true&recordMetadata%5B%5D=commentCount'
             );
             res.json({
                 records: [
@@ -327,6 +327,7 @@ describe('record selection', function() {
                         id: 'recordA',
                         fields: {Name: 'Rebecca'},
                         createdTime: '2020-04-20T16:20:00.000Z',
+                        commentCount: 0,
                     },
                 ],
                 offset: 'offsetABC',
@@ -340,11 +341,14 @@ describe('record selection', function() {
                 maxRecords: 50,
                 sort: [{field: 'Name', direction: 'desc'}],
                 cellFormat: 'json',
+                returnFieldsByFieldId: true,
+                recordMetadata: ['commentCount'],
             })
             .eachPage(function page(records) {
                 records.forEach(function(record) {
                     expect(record.id).toBe('recordA');
                     expect(record.get('Name')).toBe('Rebecca');
+                    expect(record.commentCount).toBe(0);
                 });
                 done();
             });
@@ -540,6 +544,83 @@ describe('record selection', function() {
                 expect(err.statusCode).toBe(402);
                 expect(err.message).toBe('foo bar');
                 expect(result).toBeNull();
+                done();
+            });
+    });
+
+    it('uses POST listRows endpoint when params exceed 15k characters', function(done) {
+        // Mock formula that is 15000 characters in length
+        const longFormula = [...Array(657)]
+            .map((_, i) => {
+                return `NOT({Name} = '${i}') & `;
+            })
+            .join();
+
+        expect(longFormula.length).toBe(15000);
+
+        testExpressApp.set('handler override', function(req, res) {
+            expect(req.method).toBe('POST');
+            expect(req.url).toBe(
+                '/v0/app123/Table/listRecords?timeZone=America%2FLos_Angeles&userLocale=en-US'
+            );
+            expect(req.body.filterByFormula).toBe(longFormula);
+
+            res.json({
+                records: [
+                    {
+                        id: 'recordA',
+                        fields: {Name: 'Rebecca'},
+                        createdTime: '2020-04-20T16:20:00.000Z',
+                    },
+                ],
+                offset: 'offsetABC',
+            });
+        });
+
+        return airtable
+            .base('app123')
+            .table('Table')
+            .select({
+                filterByFormula: longFormula,
+                timeZone: 'America/Los_Angeles',
+                userLocale: 'en-US',
+            })
+            .eachPage(function page(records) {
+                records.forEach(function(record) {
+                    expect(record.id).toBe('recordA');
+                    expect(record.get('Name')).toBe('Rebecca');
+                });
+                done();
+            });
+    });
+
+    it('uses POST listRows endpoint when "post" is specified for method', function(done) {
+        testExpressApp.set('handler override', function(req, res) {
+            expect(req.method).toBe('POST');
+            expect(req.url).toBe('/v0/app123/Table/listRecords?');
+            res.json({
+                records: [
+                    {
+                        id: 'recordA',
+                        fields: {Name: 'Rebecca'},
+                        createdTime: '2020-04-20T16:20:00.000Z',
+                    },
+                ],
+                offset: 'offsetABC',
+            });
+        });
+
+        return airtable
+            .base('app123')
+            .table('Table')
+            .select({
+                method: 'post',
+            })
+            .eachPage(function page(records) {
+                records.forEach(function(record) {
+                    expect(record.id).toBe('recordA');
+                    expect(record.get('Name')).toBe('Rebecca');
+                });
                 done();
             });
     });